Rename constants.mjs to constants.js
